% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essa_ekch.R
\docType{data}
\name{essa_ekch}
\alias{essa_ekch}
\title{4D state vectors for flights between ESSA-EKCH}
\format{
\subsection{\verb{traffic state vector}}{

A data frame with 390788 rows and 6 columns:
\describe{
\item{callsign}{flight identification}
\item{icao24}{ModeS airframe address}
\item{timestamp}{date time of position report}
\item{altitude}{reported altitude in ft (traffic lib converts m to ft)}
\item{latitude}{latitude of pasition}
\item{longitude}{longitude of position}
}
}
}
\source{
PRU internal dataset, http unknown
}
\usage{
essa_ekch
}
\description{
A subset of the data extracted for a period in 2023 from OSN using the
traffic-library. These data were subset to FLTID & 4D.
}
\keyword{datasets}
