% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trj_identify_legs.R
\name{identify_trajectory_legs}
\alias{identify_trajectory_legs}
\title{Identify legs / gaps in a trajectory}
\usage{
identify_trajectory_legs(.trjs, .grp_var = .env$ICAO24, .max_gap = 3)
}
\arguments{
\item{.trjs}{tibble/dataframe of state vectors (flight identification and 4D)}

\item{.grp_var}{variable in .trjs used to separate/group flights, defaults to ICAO24}

\item{.max_gap}{in minutes; defaults to 3 minutes (e.g. low approach)}
}
\value{
augmented state vector tibble
}
\description{
Label legs (segments) of a trajectory (time ordered sequence of ICAO24
addresses). Add a colum to identify the leg.
}
\examples{
\dontrun{
state_vectors |> identify_trajectory_legs()
}
}
