% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pru3di-package.R
\docType{package}
\name{pru3di-package}
\alias{pru3di}
\alias{pru3di-package}
\title{pru3di: An opinionated hack to implement the Borealis use-case study on 3Di}
\description{
An opinionated implementation of the 3Di algorithm. The use case serves also to allow me to explore the development (and it workflow) of R packages. This work-in-progress will ultimately be transferred to a more useful and stable version.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rainer-rq-koelle/pru3di}
  \item \url{https://rainer-rq-koelle.github.io/pru3di/}
  \item Report bugs at \url{https://github.com/rainer-rq-koelle/pru3di/issues}
}

}
\author{
\strong{Maintainer}: Rainer Koelle \email{rainer.koelle@eurocontrol.int} (\href{https://orcid.org/0000-0003-2810-1404}{ORCID})

}
\keyword{internal}
