% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trj_standardise_names_trafficlib.R
\name{standardise_traj_from_trafficlib}
\alias{standardise_traj_from_trafficlib}
\title{Standardise trajectory state vectors from traffic library.}
\usage{
standardise_traj_from_trafficlib(.svs_tfclib)
}
\arguments{
\item{.svs_tfclib}{tibble with state vectors for flights}
}
\value{
standard named tibble
}
\description{
Ensure consistent naming conventions and data types
}
\examples{
\dontrun{
state_vector_sample |> standardise_traj_from_trafficlib()
}
}
