% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xhello.R
\name{xhello}
\alias{xhello}
\title{A friendly greeting for a person we know by name}
\usage{
xhello(.name)
}
\arguments{
\item{.name}{the name of the person we want to greet}
}
\value{
string comprising "Greeting + name provided
}
\description{
A friendly greeting for a person we know by name
}
\examples{
xhello("Stranger")
}
